Moved the encoding method from the agent to the environment. Allowed exportGIF to be passed a direct ndarray. Still need to add a gameMemory method...
