+++ Snake/SnakeAgent.py
@@ -124,108 +124,3 @@ class SnakeAgent:
         self.direction = newDirection  # Set to new direction...
         self.currentFrame = newState
         return reward, self.gameOver
-
-    def encodeCurrentState(self):
-        """
-        Primarily internal method. It will take the current
-        state of the snake, and encode it according to our rules.
-        It will use the most recent location of the snake in
-        the environment variables.
-        :return: The state coded as an 11-bit string:
-            - Is there immediate danger in front, left,
-            or right of the snake?
-            - The direction of the fruit (up, down, left,
-            right). From a top-down perspective. More
-            than one is possible.
-            - The direction of the snake (up, down, left,
-            right)
-        Thus, the coding is [danger ==> 'FLR']
-        [fruit direction ==> 'UDLR']
-        [snake direction ==> 'UDLR'] (mutually exclusive)
-        """
-        currState = self.currentFrame[-1]
-        directionCode = {
-            'U': '1000',
-            'D': '0100',
-            'L': '0010',
-            'R': '0001'
-        }
-        coding = ''
-        # For immediate danger, we look at the snake head, and see
-        # if either the edge of the board or a snake body part is
-        # next to it. The array is in FLR order.
-        head = currState[-1]
-        if self.direction == 'U':
-            proximity = [
-                (head[0] - 1, head[1]),
-                (head[0], head[1] - 1),
-                (head[0], head[1] + 1)
-            ]
-        elif self.direction == 'D':
-            proximity = [
-                (head[0] + 1, head[1]),
-                (head[0], head[1] + 1),
-                (head[0], head[1] - 1)
-            ]
-        elif self.direction == 'L':
-            proximity = [
-                (head[0], head[1] - 1),
-                (head[0] + 1, head[1]),
-                (head[0] - 1, head[1])
-            ]
-        else:
-            proximity = [
-                (head[0], head[1] + 1),
-                (head[0] - 1, head[1]),
-                (head[0] + 1, head[1])
-            ]
-        # Lotta stuff going on here:
-        #   Check if each location is in the snake body
-        #   or off the board. Convert the Trues and Falses
-        # into a bit string we can directly attach to our coding.
-        dangers = ((r, c) in currState or not (0 <= r < self.env.boardSize and 0 <= c < self.env.boardSize)
-                   for r, c in proximity)
-        coding += ''.join(map(lambda x: str(int(x)), dangers))
-
-        # Now the fruit location. The fruit can't be both above and
-        # below the snake, so append in pairs...
-        fruitR, fruitC = self.env.fruitLocs[-1]
-        if head[0] > fruitR:
-            coding += '10'
-        elif head[0] < fruitR:
-            coding += '01'
-        else:
-            coding += '00'  # The fruit is on the same row
-        # Left/right
-        if head[1] > fruitC:
-            coding += '10'
-        elif head[1] < fruitC:
-            coding += '01'
-        else:
-            coding += '00'
-
-        # Now the direction of the snake...Straightforward
-        coding += directionCode[self.direction]
-        return coding
-
-    def getGameMemory(self):
-        """
-        Converts all the frames of the game into a format
-        for Q-learning. Ideally should be called when
-        the game is over, but since we are saving
-        the encoded states as we go along, you can
-        call it whenever....
-        :return: A list of tuples in the form
-        (state, action, reward, nextState, gameOver).
-        """
-        # The length of the actionRewards list will always be
-        # one less than the length of the states. Thus,
-        # because of what we're keeping track of,
-        # this is a simple loop...
-        memory = []
-        for i in range(len(self.actionsRewards)):
-            currentState = self.encodedStates[i]
-            nextState = self.encodedStates[i + 1]
-            _, action, reward, gameOver = self.actionsRewards[i]
-            memory.append((currentState, action, reward, nextState, gameOver))
-        return memory
