+++ Snake/SnakeQTable.py
@@ -13,7 +13,9 @@ is a pretty straightforward 2^11 by 3 table.
 import numpy as np
 from .SnakeAgent import SnakeAgent
 from .SnakeEnv import SnakeGame
+from .utils import *
 
+import numpy as np
 
 class SnakeQTable:
     def __init__(self, boardSize=20):
@@ -26,8 +28,8 @@ class SnakeQTable:
         self.minEpsilon = 0.01
         self.gamma = 0.9
         self.Qtable = np.random.random((2 ** 11, 3))
-        self.env = SnakeGame(boardSize=boardSize)
-        self.agent = SnakeAgent(environment=self.env)
+        self.agent = SnakeAgent()
+        self.env = SnakeGame(snakeAgent=self.agent, boardSize=boardSize)
 
     def playGame(self, makeGif=False, random=True):
         """
@@ -39,10 +41,16 @@ class SnakeQTable:
         :return: The encoded game memory.
         """
         self.gamesPlayed += 1
+        stateCounts = 1
         # Reset the agent
-        self.agent.reset()
-        while not self.agent.gameOver and len(self.agent.snakeFrames) < 10000:
-            currentEncodedState = self.agent.encodeCurrentState()
+        currentState = self.env.reset()
+        # Only save the frames if we're making a GIF...
+        allSnakeStates = []
+        if makeGif:
+            allSnakeStates = [produceBoardFrame(snakeGameState=currentState)]  # One frame with the first state...
+        gameOver = False
+        while not gameOver and stateCounts < 10000:
+            currentEncodedState = self.env.encodeCurrentState()
             # With an epsilon% chance, choose
             # a random action. Otherwise, choose
             # the action with the largest Q-value.
@@ -52,14 +60,16 @@ class SnakeQTable:
                 row = int(currentEncodedState, 2)
                 rowData = self.Qtable[row]
                 action = self.agent.actionList[np.argmax(rowData)]
-            reward, gameOver = self.agent.makeMove(action)
+            currentState, reward, gameOver = self.env.stepForward(action)
+            if makeGif:
+                allSnakeStates.append(produceBoardFrame(currentState))
+            stateCounts += 1
         # Game is over, so return the game memory...
         if self.agent.score > self.maxScore:
             self.maxScore = self.agent.score
-            self.env.exportGIF(f'BestGame.gif', self.agent.snakeFrames, scale=15)
         if makeGif:
-            self.env.exportGIF(f'Game{self.gamesPlayed}.gif', self.agent.snakeFrames, scale=15)
-            print(f'Game {self.gamesPlayed} scored {self.agent.score}! (Total states: {len(self.agent.snakeFrames)}, '
+            exportGIF(frames=allSnakeStates, filename=f'Game{self.gamesPlayed}.gif', scale=15)
+            print(f'Game {self.gamesPlayed} scored {self.agent.score}! '
                   f'Best Score: {self.maxScore})')
         return self.agent.getGameMemory()
 
@@ -88,7 +98,7 @@ class SnakeQTable:
 
 if __name__ == '__main__':
     qtableObj = SnakeQTable()
-    gamesToPlay = 10000
+    gamesToPlay = 5
     for game in range(1, gamesToPlay + 1):
         print(f'\rGame {game}...', end='')
         gameMem = qtableObj.playGame()
